requirements.txt
discord.py
keyboard